Convert SCSS to CSS to work with Administrate 1.0.0.beta3 (#88)
Fixes #85 

Convert SCSS to CSS so that `sassc` is not a dependency any more, since it
won't be part of Administrate v1.

In addition to the usual installation steps, I also had to add the following
to my app/assets/config/manifest.js

```
//= link administrate-field-nested_has_many/application.css
//= link administrate-field-nested_has_many/application.js
```

I tested this change on our production deployment of 1.0.0.beta3 and it works
fine.